Updated image.py
Changes:
- renamed image_generator.py as image.py for simplicity;
- implemented Generator._new_image_from_template();
- implemented Generator._save_image_as();
- other minor improvements.
